chore: move ScrollToTop out of components
