+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import React, { Suspense } from 'react'
 import { Switch, Route, BrowserRouter, Redirect } from 'react-router-dom'
 import GoogleAnalytics from 'src/pages/common/GoogleAnalytics'
 import { NotFoundPage } from './NotFound/NotFound'
-import ScrollToTop from './../components/ScrollToTop/ScrollToTop'
+import ScrollToTop from '../common/ScrollToTop'
 import Header from './common/Header/Header'
 import { SWUpdateNotification } from 'src/pages/common/SWUpdateNotification/SWUpdateNotification'
 import Main from 'src/pages/common/Layout/Main'
